// src/app/demo/api/task.ts
export interface Task {
 // assignedEquipmentId: null;
  //assignedEmployeeId: null;
  /*  id?: number;
   /* name: string; 
   description: string;
   start_date: string;
   end_date: string;
   project_id: number; */
  id?: number; 
  description: string; 
  start_date: string; 
  end_date: string; 
  due_date: string; 
  project_id: number; 
  status: string; 


}